feat(tokens): add token search modal and token management hook
